API re-structure (WIP)
