+++ integration/nevermined/AgreementStoreManager.test.ts
@@ -45,7 +45,7 @@ describe('Agreement Store Manager', () => {
         await account2.requestTokens(
             +ddo.getPriceByService() * 10 ** -(await nevermined.keeper.token.decimals())
         )
-        const agreementId = await nevermined.assets.order(ddo.id, 'access', account2)
+        const agreementId = await nevermined.assets.order(ddo.id, account2)
 
         // wait for the graph to pickup the event
         await sleep(3000)
