+++ integration/nevermined/NFT.test.ts
@@ -11,7 +11,7 @@ import {
     getRoyaltyAttributes,
     RoyaltyAttributes,
     RoyaltyKind
-} from '../../src/nevermined/Assets'
+} from '../../src/nevermined/api/AssetsApi'
 
 describe('Nfts operations', () => {
     let nevermined: Nevermined
