+++ src/nevermined/api/ConditionsApi.ts
@@ -1,0 +1,5 @@
+import { NVMBaseApi } from "./NVMBaseApi";
+
+
+export class ConditionsApi extends NVMBaseApi {
+}
