+++ src/nevermined/api/NVMBaseApi.ts
@@ -1,0 +1,8 @@
+import { Instantiable } from "../../Instantiable.abstract"
+
+export abstract class NVMBaseApi extends Instantiable {
+
+
+
+
+}
