+++ src/nevermined/api/UtilsApi.ts
@@ -1,0 +1,5 @@
+import { NVMBaseApi } from "./NVMBaseApi";
+
+
+export class UtilsApi extends NVMBaseApi {
+}
