+++ src/nevermined/api/nfts/PoapApi.ts
@@ -1,17 +1,17 @@
-import { Account, Config, Nevermined } from '../..';
-import { InstantiableConfig } from '../../Instantiable.abstract';
-import ContractHandler from '../../keeper/ContractHandler';
-import Nft721Contract from '../../keeper/contracts/Nft721Contract';
+import { Account, Config, Nevermined } from '../../..';
+import { InstantiableConfig } from '../../../Instantiable.abstract';
+import ContractHandler from '../../../keeper/ContractHandler';
+import Nft721Contract from '../../../keeper/contracts/Nft721Contract';
 import { NFT721Api } from './NFT721Api'
 
-export default class SoulBoundNFTApi extends NFT721Api {  
+export default class PoapNFTApi extends NFT721Api {  
 
     public static async getInstanceUsingABI(
         config: InstantiableConfig,
         nftContractAddress: string,
         solidityABI: any
-    ): Promise<SoulBoundNFTApi> {
-        const nft = new SoulBoundNFTApi()
+    ): Promise<PoapNFTApi> {
+        const nft = new PoapNFTApi()
         nft.setInstanceConfig(config)
 
         nft.nftContract = await Nft721Contract.getInstanceUsingABI(
@@ -27,12 +27,12 @@ export default class SoulBoundNFTApi extends NFT721Api {
         contractABI: any,
         from: Account,
         args: string[] = []
-    ): Promise<SoulBoundNFTApi> {
+    ): Promise<PoapNFTApi> {
                 
         const { instanceConfig } = (await Nevermined.getInstance(config)) as any
         const contractHandler = new ContractHandler(instanceConfig)
         const nftContract = await contractHandler.deployAbi(contractABI, from, args)
-        return SoulBoundNFTApi.getInstanceUsingABI(instanceConfig, nftContract.address, contractABI)
+        return PoapNFTApi.getInstanceUsingABI(instanceConfig, nftContract.address, contractABI)
     }
 
 }
