+++ src/nevermined/api/nfts/SubscriptionNFTApi.ts
@@ -1,17 +1,17 @@
-import { Account, Config, Nevermined } from '../..';
-import { InstantiableConfig } from '../../Instantiable.abstract';
-import ContractHandler from '../../keeper/ContractHandler';
-import Nft721Contract from '../../keeper/contracts/Nft721Contract';
+import { Account, Config, Nevermined } from '../../..';
+import { InstantiableConfig } from '../../../Instantiable.abstract';
+import ContractHandler from '../../../keeper/ContractHandler';
+import Nft721Contract from '../../../keeper/contracts/Nft721Contract';
 import { NFT721Api } from './NFT721Api'
 
-export default class PoapNFTApi extends NFT721Api {  
+export default class SubscriptionNFTApi extends NFT721Api {  
 
     public static async getInstanceUsingABI(
         config: InstantiableConfig,
         nftContractAddress: string,
         solidityABI: any
-    ): Promise<PoapNFTApi> {
-        const nft = new PoapNFTApi()
+    ): Promise<SubscriptionNFTApi> {
+        const nft = new SubscriptionNFTApi()
         nft.setInstanceConfig(config)
 
         nft.nftContract = await Nft721Contract.getInstanceUsingABI(
@@ -27,12 +27,12 @@ export default class PoapNFTApi extends NFT721Api {
         contractABI: any,
         from: Account,
         args: string[] = []
-    ): Promise<PoapNFTApi> {
+    ): Promise<SubscriptionNFTApi> {
                 
         const { instanceConfig } = (await Nevermined.getInstance(config)) as any
         const contractHandler = new ContractHandler(instanceConfig)
         const nftContract = await contractHandler.deployAbi(contractABI, from, args)
-        return PoapNFTApi.getInstanceUsingABI(instanceConfig, nftContract.address, contractABI)
+        return SubscriptionNFTApi.getInstanceUsingABI(instanceConfig, nftContract.address, contractABI)
     }
 
 }
