+++ src/node/NeverminedNode.ts
@@ -7,6 +7,9 @@ import { NeverminedNodeError, HttpError } from '../errors'
 import { ServiceType } from '../ddo/Service'
 import BigNumber from '../utils/BigNumber'
 import { ERCType } from '../models/NFTAttributes'
+import { DDO } from '../ddo/DDO'
+import { PublishMetadata } from '../nevermined/api/AssetsApi'
+import { ImmutableBackends } from '../ddo/NvmConfig'
 
 const apiPath = '/api/v1/node/services'
 
@@ -447,6 +450,34 @@ export class NeverminedNode extends Instantiable {
 
         const jsonPayload = await response.json()
         return jsonPayload.access_token
+    }
+
+    public async publishImmutableContent(
+        ddo: DDO,
+        publishMetadata: PublishMetadata = PublishMetadata.IPFS
+    ): Promise<{url: string, backend: ImmutableBackends}> {
+        let url, backend = undefined
+
+        if (publishMetadata === PublishMetadata.Filecoin) {
+            this.logger.log('Publishing metadata to Filecoin')                    
+            ;({ url } = await this.nevermined.node.uploadContent(
+                JSON.stringify(ddo), 
+                false, 
+                NodeUploadBackends.Filecoin
+                ))
+            backend = 'filecoin'
+            
+        } else if (publishMetadata === PublishMetadata.IPFS) {
+            this.logger.log('Publishing metadata to IPFS')                    
+            ;({ url: url } = await this.nevermined.node.uploadContent(
+                JSON.stringify(ddo), 
+                false, 
+                NodeUploadBackends.IPFS
+                ))
+            backend = 'ipfs'
+
+        }
+        return { url, backend }
     }
 
     public async uploadContent(data: ReadStream | string, encrypt?: boolean, backend: NodeUploadBackends = NodeUploadBackends.Filecoin): Promise<any> {
